@@ -18,35 +18,36 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~14.3.0",
-    "@angular/cdk": "^14.2.7",
-    "@angular/common": "~14.3.0",
-    "@angular/compiler": "~14.3.0",
-    "@angular/core": "~14.3.0",
-    "@angular/forms": "~14.3.0",
-    "@angular/material": "^14.2.7",
-    "@angular/platform-browser": "~14.3.0",
-    "@angular/platform-browser-dynamic": "~14.3.0",
-    "@angular/router": "~14.3.0",
-    "bootstrap": "^3.3.7",
+    "@angular/animations": "~15.2.10",
+    "@angular/cdk": "^15.2.9",
+    "@angular/common": "~15.2.10",
+    "@angular/compiler": "~15.2.10",
+    "@angular/core": "~15.2.10",
+    "@angular/forms": "~15.2.10",
+    "@angular/material": "^15.2.9",
+    "@angular/platform-browser": "~15.2.10",
+    "@angular/platform-browser-dynamic": "~15.2.10",
+    "@angular/router": "~15.2.10",
+    "bootstrap": "5.3.3",
     "rxjs": "~6.6.0",
     "tslib": "^2.0.0",
+    "uuid": "^9.0.1",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~14.2.13",
-    "@angular-eslint/builder": "14.4.0",
-    "@angular-eslint/eslint-plugin": "14.4.0",
-    "@angular-eslint/eslint-plugin-template": "14.4.0",
-    "@angular-eslint/schematics": "14.4.0",
-    "@angular-eslint/template-parser": "14.4.0",
-    "@angular/cli": "~14.2.13",
-    "@angular/compiler-cli": "~14.3.0",
+    "@angular-devkit/build-angular": "~15.2.11",
+    "@angular-eslint/builder": "15.2.1",
+    "@angular-eslint/eslint-plugin": "15.2.1",
+    "@angular-eslint/eslint-plugin-template": "15.2.1",
+    "@angular-eslint/schematics": "15.2.1",
+    "@angular-eslint/template-parser": "15.2.1",
+    "@angular/cli": "~15.2.11",
+    "@angular/compiler-cli": "~15.2.10",
     "@types/jasmine": "~3.6.0",
     "@types/node": "^12.11.1",
-    "@typescript-eslint/eslint-plugin": "5.3.0",
-    "@typescript-eslint/parser": "5.3.0",
-    "eslint": "^8.2.0",
+    "@typescript-eslint/eslint-plugin": "^5.43.0",
+    "@typescript-eslint/parser": "^5.43.0",
+    "eslint": "^8.28.0",
     "eslint-plugin-import": "2.25.2",
     "eslint-plugin-jsdoc": "latest",
     "eslint-plugin-prefer-arrow": "latest",
@@ -58,6 +59,6 @@
     "karma-jasmine": "~4.0.2",
     "karma-jasmine-html-reporter": "^1.7.0",
     "ts-node": "~8.3.0",
-    "typescript": "~4.6.4"
+    "typescript": "~4.9.5"
   }
 }
