@@ -2,24 +2,18 @@ import { BrowserModule } from '@angular/platform-browser'
 import { NgModule } from '@angular/core'
 
 import { BackgroundComponent } from './components/background/background.component'
-import { MatTabsModule } from '@angular/material/tabs'
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations'
 import { AdmissionComponent } from './components/admission/admission.component'
 import { HomePatientComponent } from './components/home-patient/home-patient.component'
 import { HeaderComponent } from './components/header/header.component'
 import { LogInComponent } from './components/log-in/log-in.component'
-import { MatFormFieldModule } from '@angular/material/form-field'
 import { MatIconModule } from '@angular/material/icon'
 import { FormsModule, ReactiveFormsModule } from '@angular/forms'
-import { MatCardModule } from '@angular/material/card'
 import { LoginService } from './components/log-in/login.service'
 import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http'
 import { AppConfigService } from './core/app-config.service'
-import { MatSnackBarModule } from '@angular/material/snack-bar'
-import { CommonModule } from '@angular/common'
-import { MatInputModule } from '@angular/material/input'
+import { CommonModule, NgOptimizedImage } from '@angular/common'
 import { HomePersonalComponent } from './components/home-personal/home-personal.component'
-import { MatTooltipModule } from '@angular/material/tooltip'
 import { PatientInfoViewComponent } from './components/patient-info-view/patient-info-view.component'
 import { PersonalInfoViewComponent } from './components/personal-info-view/personal-info-view.component'
 import { AppComponent } from './app.component'
@@ -27,15 +21,11 @@ import { AppRoutingModule } from './app-routing.module'
 import { DiagnosisListComponent } from './components/personal-info-view/diagnosis-list/diagnosis-list.component'
 import { PersonListComponent } from './components/personal-info-view/person-list/person-list.component'
 import { ResetPasswordComponent } from './components/log-in/reset-password/reset-password.component'
-import { MatTableModule } from '@angular/material/table'
 import { DiseaseListComponent } from './components/personal-info-view/disease-list/disease-list.component'
-import { MatButtonModule } from '@angular/material/button'
 import { MatStepperModule } from '@angular/material/stepper'
-import { MatSelectModule } from '@angular/material/select'
 import { AllergyListComponent } from './components/personal-info-view/allergy-list/allergy-list.component'
 import { MedicationInTakeListComponent } from './components/personal-info-view/medication-in-take-list/medication-in-take-list.component'
 import { MyProfileModalComponent } from './components/home-patient/my-profile-modal/my-profile-modal.component'
-import { MatDialogModule } from '@angular/material/dialog'
 import { MatDividerModule } from '@angular/material/divider'
 import { AllergyListPatientComponent } from './components/patient-info-view/allergy-list-Patient/allergy-list-patient.component'
 import { DiagnosisListPatientComponent } from './components/patient-info-view/diagnosis-list-patient/diagnosis-list-patient.component'
@@ -53,8 +43,18 @@ import { StartNewRegistrationModalComponent } from './shared/dialogs/start-new-r
 import { PersonViewComponent } from './components/patient-info-view/person-view/person-view.component'
 import { MatDatepickerModule } from '@angular/material/datepicker'
 import { MatNativeDateModule } from '@angular/material/core'
-import { MatSlideToggleModule } from '@angular/material/slide-toggle'
 import { MatButtonToggleModule } from '@angular/material/button-toggle'
+import { MatFormFieldModule } from '@angular/material/form-field'
+import { MatInputModule } from '@angular/material/input'
+import { MatCardModule } from '@angular/material/card'
+import { MatSnackBarModule } from '@angular/material/snack-bar'
+import { MatTooltipModule } from '@angular/material/tooltip'
+import { MatTabsModule } from '@angular/material/tabs'
+import { MatTableModule } from '@angular/material/table'
+import { MatButtonModule } from '@angular/material/button'
+import { MatDialogModule } from '@angular/material/dialog'
+import { MatSelectModule } from '@angular/material/select'
+import { MatSlideToggleModule } from '@angular/material/slide-toggle'
 import { MatAutocompleteModule } from '@angular/material/autocomplete'
 
 
@@ -96,14 +96,14 @@ const angularMaterialModules = [
         StartNewRegistrationModalComponent,
         PersonViewComponent
     ],
-    imports: [
-        angularMaterialModules,
-        BrowserModule,
-        HttpClientModule,
-        AppRoutingModule,
-        BrowserAnimationsModule,
-        FormsModule, ReactiveFormsModule, CommonModule
-    ],
+  imports: [
+    angularMaterialModules,
+    BrowserModule,
+    HttpClientModule,
+    AppRoutingModule,
+    BrowserAnimationsModule,
+    FormsModule, ReactiveFormsModule, CommonModule, NgOptimizedImage
+  ],
     providers: [
         {
             provide: HTTP_INTERCEPTORS,
