@@ -1,12 +1,12 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing'
 
 import { HeaderComponent } from './header.component'
-import { MatSnackBarModule } from '@angular/material/snack-bar'
 import { RouterTestingModule } from '@angular/router/testing'
 import { HttpClientTestingModule } from '@angular/common/http/testing'
 import { LoginService } from '../log-in/login.service'
 import { SessionService } from '../../core/authentification-and-authority/session.service'
 import { AppConfigService } from '../../core/app-config.service'
+import { MatSnackBarModule } from '@angular/material/snack-bar'
 
 describe('HeaderComponent', () => {
   let component: HeaderComponent
@@ -19,8 +19,8 @@ describe('HeaderComponent', () => {
         RouterTestingModule, HttpClientTestingModule, MatSnackBarModule
       ],
       providers: [
-      LoginService, SessionService, AppConfigService
-    ],
+        LoginService, SessionService, AppConfigService
+      ],
     })
       .compileComponents()
   })
