@@ -6,9 +6,9 @@ import { HttpClientTestingModule } from '@angular/common/http/testing'
 import { LoginService } from './login.service'
 import { SessionService } from '../../core/authentification-and-authority/session.service'
 import { AppConfigService } from '../../core/app-config.service'
-import { MatCardModule } from '@angular/material/card'
-import { MatTabsModule } from '@angular/material/tabs'
-import { MatSnackBarModule } from '@angular/material/snack-bar'
+import { MatLegacyCardModule as MatCardModule } from '@angular/material/legacy-card'
+import { MatLegacyTabsModule as MatTabsModule } from '@angular/material/legacy-tabs'
+import { MatLegacySnackBarModule as MatSnackBarModule } from '@angular/material/legacy-snack-bar'
 import { FormsModule, ReactiveFormsModule } from '@angular/forms'
 import { CommonModule } from '@angular/common'
 
