@@ -4,10 +4,10 @@ import { ResetPasswordComponent } from './reset-password.component'
 import { RouterTestingModule } from '@angular/router/testing'
 import { HttpClientTestingModule } from '@angular/common/http/testing'
 import { CommonModule } from '@angular/common'
+import { FormsModule, ReactiveFormsModule } from '@angular/forms'
 import { MatCardModule } from '@angular/material/card'
 import { MatTabsModule } from '@angular/material/tabs'
 import { MatSnackBarModule } from '@angular/material/snack-bar'
-import { FormsModule, ReactiveFormsModule } from '@angular/forms'
 
 describe('ResetPasswordComponent', () => {
   let component: ResetPasswordComponent
