@@ -1,14 +1,14 @@
 import { Component, Input, OnChanges, OnInit, SimpleChange } from '@angular/core'
 import { v4 as uuid } from 'uuid'
 import { DiseaseService } from '../../services/disease.service'
-import { MatSnackBar } from '@angular/material/snack-bar'
-import { MatDialog } from '@angular/material/dialog'
 import { SessionService } from '../../../core/authentification-and-authority/session.service'
 import { IDisease } from '../../../model/disease.interface'
 import { DiseaseModalComponent } from '../../../shared/dialogs/disease-modal/disease-modal.component'
 import { DeleteConfirmationComponent } from '../../../shared/dialogs/delete-confirmation-modal/delete-confirmation.component'
 import { IDeleteConfirmation } from '../../../model/delete-confirmation.interface'
 import { IPerson } from '../../../model/person.interface'
+import { MatSnackBar } from '@angular/material/snack-bar'
+import { MatDialog, MatDialogRef } from '@angular/material/dialog'
 
 @Component({
   selector: 'app-disease-list-patient',
@@ -16,7 +16,7 @@ import { IPerson } from '../../../model/person.interface'
   styleUrls: ['./disease-list-patient.component.css']
 })
 export class DiseaseListPatientComponent implements OnInit, OnChanges {
-  displayedColumns: string[] = ['#', 'preExistingIllness', 'undergoneSurgery', 'surgeryReason', 'action']
+  displayedColumns: Array<string> = ['#', 'preExistingIllness', 'undergoneSurgery', 'surgeryReason', 'action']
   diseasesList: Array<IDisease> = []
   fromInit: boolean = true
   @Input() currentUser: IPerson = <IPerson>{}
@@ -26,25 +26,21 @@ export class DiseaseListPatientComponent implements OnInit, OnChanges {
     private snackBar: MatSnackBar,
     private sessionService: SessionService,
     public dialog: MatDialog,
-  ) { }
+  ) {
+  }
 
   ngOnInit(): void {
     this.listPersonDiseases()
   }
 
   ngOnChanges(changes: { [propertyName: string]: SimpleChange }): void {
-    console.log(changes)
-
-    if (changes.currentUser && this.currentUser && this.currentUser.id) {
-      if (!this.fromInit) {
-        this.listPersonDiseases()
-      }
+    if (changes.currentUser && this.currentUser && this.currentUser.id && !this.fromInit) {
+      this.listPersonDiseases()
     }
-
   }
 
   onAddNewOrEdit(disease?: IDisease): void {
-    const dialogRef = this.dialog.open(DiseaseModalComponent, {
+    const dialogRef: MatDialogRef<DiseaseModalComponent> = this.dialog.open(DiseaseModalComponent, {
       width: '750px',
       data: {update: disease, parent: 'patient', patient: this.currentUser}
     })
@@ -57,7 +53,7 @@ export class DiseaseListPatientComponent implements OnInit, OnChanges {
   }
 
   onDelete(disease: IDisease): void {
-    const dialogRef = this.dialog.open(DeleteConfirmationComponent, {
+    const dialogRef: MatDialogRef<DeleteConfirmationComponent> = this.dialog.open(DeleteConfirmationComponent, {
       disableClose: true,
       data: <IDeleteConfirmation>{entityType: 'Disease', entityName: ''}
     })
@@ -71,10 +67,10 @@ export class DiseaseListPatientComponent implements OnInit, OnChanges {
 
 
   private listDisease(): void {
-    this.diseaseService.getAll().subscribe((diseases: any) => {
+    this.diseaseService.getAll().subscribe((diseases: any): void => {
         this.diseasesList = diseases
       },
-      err => {
+      (err: Error): void => {
         console.log('Error in DiseaseListPatientComponent.listDisease()')
         console.log(err)
         this.snackBar.open('Could not fetch diseases', 'Close', {
@@ -85,10 +81,10 @@ export class DiseaseListPatientComponent implements OnInit, OnChanges {
   }
 
   private deleteDisease(diseaseId: uuid): void {
-    this.diseaseService.delete(diseaseId).subscribe(() => {
+    this.diseaseService.delete(diseaseId).subscribe((): void => {
         this.listDisease()
       },
-      err => {
+      (err: Error): void => {
         console.log('Error in DiseaseListPatientComponent.deleteDisease()')
         console.log(err)
         this.snackBar.open('Could not delete disease', 'Close', {
@@ -99,18 +95,16 @@ export class DiseaseListPatientComponent implements OnInit, OnChanges {
   }
 
   private listPersonDiseases(): void {
-    let patientId: uuid = ''
-    if (this.currentUser && this.currentUser.id) {
-      patientId = this.currentUser.id
-    } else {
-      patientId = this.sessionService.getUserId()
-    }
-    this.diseaseService.getAllByPersonId(patientId).subscribe((diseases: Array<IDisease>) => {
+    const patientId: uuid = this.currentUser && this.currentUser.id
+      ? this.currentUser.id
+      : this.sessionService.getUserId()
+
+    this.diseaseService.getAllByPersonId(patientId).subscribe((diseases: Array<IDisease>): void => {
         console.log('diseases fetched')
         this.fromInit = false
         this.diseasesList = diseases
       },
-      err => {
+      (err: Error): void => {
         console.log('Error in DiseaseListPatientComponent.listPersonDiseases()')
         console.log(err)
         this.snackBar.open('Could not fetch diseases', 'Close', {
