@@ -3,12 +3,12 @@ import { v4 as uuid } from 'uuid'
 import { IPerson } from '../../../model/person.interface'
 import { IMedication } from '../../../model/medication.interface'
 import { MedicationService } from '../../services/medication.service'
-import { MatSnackBar } from '@angular/material/snack-bar'
 import { SessionService } from '../../../core/authentification-and-authority/session.service'
-import { MatDialog } from '@angular/material/dialog'
 import { MedicationInTakeModalComponent } from '../../../shared/dialogs/medication-in-take-modal/medication-in-take-modal.component'
 import { DeleteConfirmationComponent } from '../../../shared/dialogs/delete-confirmation-modal/delete-confirmation.component'
 import { IDeleteConfirmation } from '../../../model/delete-confirmation.interface'
+import { MatSnackBar } from '@angular/material/snack-bar'
+import { MatDialog, MatDialogRef } from '@angular/material/dialog'
 
 @Component({
   selector: 'app-medication-in-take-list-patient',
@@ -16,7 +16,7 @@ import { IDeleteConfirmation } from '../../../model/delete-confirmation.interfac
   styleUrls: ['./medication-in-take-list-patient.component.css']
 })
 export class MedicationInTakeListPatientComponent implements OnInit {
-  displayedColumns: string[] = ['#', 'Bezeichnung ', 'Dosierung', 'Startdatum', 'bloodDiluent', 'action']
+  displayedColumns: Array<string> = ['#', 'Bezeichnung ', 'Dosierung', 'Startdatum', 'bloodDiluent', 'action']
   medicationInTakeList: Array<IMedication> = []
   @Input() currentUser: IPerson = <IPerson>{}
 
@@ -25,14 +25,15 @@ export class MedicationInTakeListPatientComponent implements OnInit {
     private sessionService: SessionService,
     private snackBar: MatSnackBar,
     public dialog: MatDialog,
-  ) { }
+  ) {
+  }
 
   ngOnInit(): void {
     this.listPatientMedicationInTake()
   }
 
   onAddNewOrEdit(medication?: IMedication): void {
-    const dialogRef = this.dialog.open(MedicationInTakeModalComponent, {
+    const dialogRef: MatDialogRef<MedicationInTakeModalComponent> = this.dialog.open(MedicationInTakeModalComponent, {
       width: '750px',
       data: {update: medication, parent: 'patient', patient: this.currentUser}
     })
@@ -45,7 +46,7 @@ export class MedicationInTakeListPatientComponent implements OnInit {
   }
 
   onDelete(medication: IMedication): void {
-    const dialogRef = this.dialog.open(DeleteConfirmationComponent, {
+    const dialogRef: MatDialogRef<DeleteConfirmationComponent> = this.dialog.open(DeleteConfirmationComponent, {
       disableClose: true,
       data: <IDeleteConfirmation>{entityType: 'Medikament', entityName: medication.designation}
     })
@@ -58,10 +59,10 @@ export class MedicationInTakeListPatientComponent implements OnInit {
   }
 
   private saveOnAddNew(medication: IMedication): void {
-    this.medicationService.add(medication).subscribe(() => {
+    this.medicationService.add(medication).subscribe((): void => {
         this.listPatientMedicationInTake()
       },
-      err => {
+      (err: Error): void => {
         console.log('Error in MedicationInTakeListPatientComponent.saveAddNewMedication()')
         console.log(err)
         this.snackBar.open('Could not add this medication', 'Close', {
@@ -73,10 +74,10 @@ export class MedicationInTakeListPatientComponent implements OnInit {
 
   private listPatientMedicationInTake(): void {
     this.medicationService.getAllByPatientId(this.sessionService.getUserId())
-      .subscribe((medications: Array<IPerson>) => {
+      .subscribe((medications: Array<IPerson>): void => {
           this.medicationInTakeList = medications
         },
-        err => {
+        (err: Error): void => {
           console.log('Error in MedicationInTakeListPatientComponent.listPatientMedicationInTake()')
           console.log(err)
           this.snackBar.open('Could not fetch persons', 'Close', {
@@ -87,10 +88,10 @@ export class MedicationInTakeListPatientComponent implements OnInit {
   }
 
   private deleteMedication(medicationId: uuid): void {
-    this.medicationService.delete(medicationId).subscribe(() => {
+    this.medicationService.delete(medicationId).subscribe((): void => {
         this.listPatientMedicationInTake()
       },
-      err => {
+      (err: Error): void => {
         console.log('Error in MedicationInTakeListPatientComponent.deleteMedication()')
         console.log(err)
         this.snackBar.open('Could not delete medication', 'Close', {
