@@ -4,9 +4,9 @@ import { PatientInfoViewComponent } from './patient-info-view.component'
 import { RouterTestingModule } from '@angular/router/testing'
 import { HttpClientTestingModule } from '@angular/common/http/testing'
 import { AppConfigService } from '../../core/app-config.service'
-import { MatSnackBarModule } from '@angular/material/snack-bar'
 import { SessionService } from '../../core/authentification-and-authority/session.service'
 import { PersonService } from '../services/person.service'
+import { MatSnackBarModule } from '@angular/material/snack-bar'
 
 describe('PatientInfoViewComponent', () => {
   let component: PatientInfoViewComponent
