@@ -1,13 +1,13 @@
 import { Component, Input, OnInit } from '@angular/core'
 import { v4 as uuid } from 'uuid'
 import { MedicationService } from '../../services/medication.service'
-import { MatSnackBar } from '@angular/material/snack-bar'
-import { MatDialog } from '@angular/material/dialog'
 import { DeleteConfirmationComponent } from '../../../shared/dialogs/delete-confirmation-modal/delete-confirmation.component'
 import { IDeleteConfirmation } from '../../../model/delete-confirmation.interface'
 import { IMedication } from '../../../model/medication.interface'
 import { MedicationInTakeModalComponent } from '../../../shared/dialogs/medication-in-take-modal/medication-in-take-modal.component'
 import { IPerson } from '../../../model/person.interface'
+import { MatSnackBar } from '@angular/material/snack-bar'
+import { MatDialog, MatDialogRef } from '@angular/material/dialog'
 
 @Component({
   selector: 'app-medication-in-take-list',
@@ -16,7 +16,7 @@ import { IPerson } from '../../../model/person.interface'
 })
 export class MedicationInTakeListComponent implements OnInit {
 
-  displayedColumns: string[] = ['#', 'patient', 'designation', 'dosage', 'startDate', 'action']
+  displayedColumns: Array<string> = ['#', 'patient', 'designation', 'dosage', 'startDate', 'action']
   medicationInTakeList: any
   @Input() patientsList: Array<IPerson> = []
 
@@ -24,14 +24,15 @@ export class MedicationInTakeListComponent implements OnInit {
     private medicationService: MedicationService,
     private snackBar: MatSnackBar,
     public dialog: MatDialog,
-  ) { }
+  ) {
+  }
 
   ngOnInit(): void {
     this.listMedicationInTakeList()
   }
 
   onAddNewOrEdit(medication?: IMedication): void {
-    const dialogRef = this.dialog.open(MedicationInTakeModalComponent, {
+    const dialogRef: MatDialogRef<MedicationInTakeModalComponent> = this.dialog.open(MedicationInTakeModalComponent, {
       width: '750px',
       data: {update: medication, parent: 'personal', patientsList: this.patientsList}
     })
@@ -44,7 +45,7 @@ export class MedicationInTakeListComponent implements OnInit {
   }
 
   onDelete(medication: IMedication): void {
-    const dialogRef = this.dialog.open(DeleteConfirmationComponent, {
+    const dialogRef: MatDialogRef<DeleteConfirmationComponent> = this.dialog.open(DeleteConfirmationComponent, {
       disableClose: true,
       data: <IDeleteConfirmation>{entityType: 'Medikament', entityName: medication.designation}
     })
@@ -56,27 +57,13 @@ export class MedicationInTakeListComponent implements OnInit {
     })
   }
 
-  private saveOnAddNew(medication: IMedication): void {
-    this.medicationService.add(medication).subscribe(() => {
-        this.listMedicationInTakeList()
-      },
-      err => {
-        console.log('Error in MedicationInTakeListComponent.saveAddNewMedication()')
-        console.log(err)
-        this.snackBar.open('Could not add this medication', 'Close', {
-          duration: 4000
-        })
-      }
-    )
-  }
-
   private listMedicationInTakeList(): void {
-    this.medicationService.getAll().subscribe((medications: any) => {
+    this.medicationService.getAll().subscribe((medications: any): void => {
         this.medicationInTakeList = medications
       },
-      err => {
+      (err: Error): void => {
         console.log('Error in MedicationInTakeListComponent.listMedicationInTakeList()')
-        console.log(err)
+        console.log((err))
         this.snackBar.open('Could not fetch medications in take', 'Close', {
           duration: 4000
         })
@@ -88,7 +75,7 @@ export class MedicationInTakeListComponent implements OnInit {
     this.medicationService.delete(medicationId).subscribe(() => {
         this.listMedicationInTakeList()
       },
-      err => {
+      (err: Error): void => {
         console.log('Error in MedicationInTakeListComponent.deleteMedication()')
         console.log(err)
         this.snackBar.open('Could not delete medication', 'Close', {
