@@ -1,13 +1,15 @@
 import { Component, Input, OnChanges, OnInit, SimpleChange } from '@angular/core'
 
 import { PersonService } from '../../services/person.service'
-import { MatSnackBar } from '@angular/material/snack-bar'
 import { IPerson } from '../../../model/person.interface'
 import { MyProfileModalComponent } from '../../home-patient/my-profile-modal/my-profile-modal.component'
-import { MatDialog } from '@angular/material/dialog'
 import { IDeleteConfirmation } from '../../../model/delete-confirmation.interface'
 import { DeleteConfirmationComponent } from '../../../shared/dialogs/delete-confirmation-modal/delete-confirmation.component'
-import { StartNewRegistrationModalComponent } from '../../../shared/dialogs/start-new-registration-modal/start-new-registration-modal.component'
+import {
+  StartNewRegistrationModalComponent
+} from '../../../shared/dialogs/start-new-registration-modal/start-new-registration-modal.component'
+import { MatSnackBar } from '@angular/material/snack-bar'
+import { MatDialog, MatDialogRef } from '@angular/material/dialog'
 
 @Component({
   selector: 'app-person-list',
@@ -15,9 +17,9 @@ import { StartNewRegistrationModalComponent } from '../../../shared/dialogs/star
   styleUrls: ['./person-list.component.css']
 })
 export class PersonListComponent implements OnInit, OnChanges {
-  displayedColumns: string[] = ['#', 'vorname', 'nachname', 'username', 'action']
-  personsList: any = []
-  filteredPersonsList: any = []
+  displayedColumns: Array<string> = ['#', 'vorname', 'nachname', 'username', 'action']
+  personsList: Array<IPerson> = []
+  filteredPersonsList: Array<IPerson> = []
 
   @Input() personType: string
 
@@ -26,14 +28,15 @@ export class PersonListComponent implements OnInit, OnChanges {
     private personService: PersonService,
     private snackBar: MatSnackBar,
     public dialog: MatDialog,
-  ) { }
+  ) {
+  }
 
   ngOnInit(): void {
     this.listPersons()
   }
 
   ngOnChanges(changes: { [propertyName: string]: SimpleChange }): void {
-    console.log(changes)
+
 
     if (changes.personType && this.personType && this.personsList.length > 0) {
       this.setPersonList(this.personsList)
@@ -43,7 +46,7 @@ export class PersonListComponent implements OnInit, OnChanges {
 
 
   onAddNewPerson(): void {
-    const dialogRef = this.dialog.open(StartNewRegistrationModalComponent, {
+    const dialogRef: MatDialogRef<StartNewRegistrationModalComponent> = this.dialog.open(StartNewRegistrationModalComponent, {
       data: {personType: this.personType}
     })
 
@@ -55,7 +58,7 @@ export class PersonListComponent implements OnInit, OnChanges {
   }
 
   onPersonInfo(person: IPerson): void {
-    const dialogRef = this.dialog.open(MyProfileModalComponent, {
+    const dialogRef: MatDialogRef<MyProfileModalComponent> = this.dialog.open(MyProfileModalComponent, {
       width: '750px',
       data: {person: person, personType: this.personType}
     })
@@ -68,7 +71,7 @@ export class PersonListComponent implements OnInit, OnChanges {
   }
 
   onDeletePatient(person: IPerson): void {
-    const dialogRef = this.dialog.open(DeleteConfirmationComponent, {
+    const dialogRef: MatDialogRef<DeleteConfirmationComponent> = this.dialog.open(DeleteConfirmationComponent, {
       disableClose: true,
       data: <IDeleteConfirmation>{entityType: person.type, entityName: person.firstName + ' ' + person.lastName}
     })
@@ -81,11 +84,11 @@ export class PersonListComponent implements OnInit, OnChanges {
   }
 
   private listPersons(): void {
-    this.personService.getAll().subscribe((data: any) => {
+    this.personService.getAll().subscribe((data: Array<IPerson>): void => {
         this.personsList = data
         this.setPersonList(data)
       },
-      err => {
+      (err: Error): void => {
         console.log('Error in PersonListComponent.listPersons()')
         console.log(err)
         this.snackBar.open('Could not fetch patients', 'Close', {
@@ -97,15 +100,15 @@ export class PersonListComponent implements OnInit, OnChanges {
 
   private setPersonList(data: Array<IPerson>): void {
     this.personType === 'personal'
-      ? this.filteredPersonsList = data.filter((person: IPerson) => person.type === 'personal')
-      : this.filteredPersonsList = data.filter((person: IPerson) => person.type === 'patient')
+      ? this.filteredPersonsList = data.filter((person: IPerson): boolean => person.type === 'personal')
+      : this.filteredPersonsList = data.filter((person: IPerson): boolean => person.type === 'patient')
   }
 
   private deletePerson(personId: string): void {
-    this.personService.delete(personId).subscribe(() => {
+    this.personService.delete(personId).subscribe((): void => {
         this.listPersons()
       },
-      err => {
+      (err: Error): void => {
         console.log('Error in PersonListComponent.deletePerson()')
         console.log(err)
         this.snackBar.open('Could not delete user', 'Close', {
