@@ -3,11 +3,11 @@ import { ComponentFixture, TestBed } from '@angular/core/testing'
 import { PersonalInfoViewComponent } from './personal-info-view.component'
 import { RouterTestingModule } from '@angular/router/testing'
 import { HttpClientTestingModule } from '@angular/common/http/testing'
-import { MatTableModule } from '@angular/material/table'
 import { PersonListComponent } from './person-list/person-list.component'
 import { DiagnosisListComponent } from './diagnosis-list/diagnosis-list.component'
 import { AppConfigService } from '../../core/app-config.service'
 import { PersonService } from '../services/person.service'
+import { MatTableModule } from '@angular/material/table'
 import { MatSnackBarModule } from '@angular/material/snack-bar'
 
 describe('PersonalInfoViewComponent', () => {
