@@ -1,9 +1,9 @@
 import { Component, Inject, OnInit } from '@angular/core'
-import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog'
+import { MatLegacyDialogRef as MatDialogRef, MAT_LEGACY_DIALOG_DATA as MAT_DIALOG_DATA } from '@angular/material/legacy-dialog'
 import { IAllergy, IAllergyTO, IPerson } from '../../../model/person.interface'
 import { AllergyService } from '../../../components/services/allergy.service'
 import { FormBuilder, FormGroup, Validators } from '@angular/forms'
-import { MatSnackBar } from '@angular/material/snack-bar'
+import { MatLegacySnackBar as MatSnackBar } from '@angular/material/legacy-snack-bar'
 import { PersonService } from '../../../components/services/person.service'
 import { allergyValues } from '../../constante'
 import { FilterService } from '../../../core/filter.service'
