@@ -1,10 +1,10 @@
-import { MatSnackBar } from '@angular/material/snack-bar'
+import { MatLegacySnackBar as MatSnackBar } from '@angular/material/legacy-snack-bar'
 import { Component, Inject, OnInit } from '@angular/core'
 import { IDisease, IDiseaseTO } from '../../../model/disease.interface'
 import { IPerson } from '../../../model/person.interface'
 import { PersonService } from '../../../components/services/person.service'
 import { FormBuilder, FormGroup, Validators } from '@angular/forms'
-import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog'
+import { MatLegacyDialogRef as MatDialogRef, MAT_LEGACY_DIALOG_DATA as MAT_DIALOG_DATA } from '@angular/material/legacy-dialog'
 import { DiseaseService } from '../../../components/services/disease.service'
 import { IIllness } from '../../../model/illness.interface'
 import { FilterService } from '../../../core/filter.service'
