@@ -1,8 +1,8 @@
 import { IPerson, IPersonTO } from '../../../model/person.interface'
 import { Component, Inject, OnInit } from '@angular/core'
 import { Router } from '@angular/router'
-import { MatSnackBar } from '@angular/material/snack-bar'
-import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog'
+import { MatLegacySnackBar as MatSnackBar } from '@angular/material/legacy-snack-bar'
+import { MatLegacyDialogRef as MatDialogRef, MAT_LEGACY_DIALOG_DATA as MAT_DIALOG_DATA } from '@angular/material/legacy-dialog'
 import { FormBuilder, FormGroup, Validators } from '@angular/forms'
 import { PersonService } from '../../../components/services/person.service'
 import { MyProfileModalComponent } from '../../../components/home-patient/my-profile-modal/my-profile-modal.component'
